Fix dependencies and refactor copypasta
